Sync PF-build.sh and MK404-build.sh with MK3_3.11.1 branch fix typo

+++ PF-build.sh
@@ -56,7 +56,7 @@
 #   Some may argue that this is only used by a script, BUT as soon someone accidentally or on purpose starts Arduino IDE
 #   it will use the default Arduino IDE folders and so can corrupt the build environment.
 #
-# Version: 2.0.2-Build_69
+# Version: 2.0.2-Build_80
 # Change log:
 # 12 Jan 2019, 3d-gussner, Fixed "compiler.c.elf.flags=-w -Os -Wl,-u,vfprintf -lprintf_flt -lm -Wl,--gc-sections" in 'platform.txt'
 # 16 Jan 2019, 3d-gussner, Build_2, Added development check to modify 'Configuration.h' to prevent unwanted LCD messages that Firmware is unknown
@@ -165,14 +165,27 @@
 # 23 Jun 2021, 3d-gussner, Improve MK404 usage
 # 24 Jun 2021, 3d-gussner, Fix MK404 user interaction not to show if compiling 'All' variants
 # 24 Jun 2021, 3d-gussner, MK404 is only supported on Linux at this moment.
+# 03 Jan 2022, 3d-gussner, Remove calling lang-community.sh as not needed anymore
+# 21 Jan 2022, 3d-gussner, Sort variants
+#                          Add Arduino 1.8.19 as an option
+# 25 Jan 2022, 3d-gussner, Allow upper and lower case for MK404
+# 09 Feb 2022, 3d-gussner, Add community language firmware files for MK2.5/S
+#                          Add selection of language in MK404 for MK2.5/S
+# 10 Feb 2022, 3d-gussner, Add SRCDIR for compatibility with build server
+# 13 Feb 2022, leptun    , Fix -o for "Restoring" messages after failure
 # 24 Feb 2022, 3d-gussner, Change to Arduino IDE 1.8.19 and Arduino boards 1.0.5
 #                          Fix DEV_STATUS to set correctly on RC/BETA/ALPHA/DEVEL
 #                          Fix atmegaMK404 Board mem and flash modifications
 #                          Limit atmegaMK404 boards mem to 8,16,32
-# 20 Jun 2022, 3d-gussner, Change to Ardunio_boards v 1.0.5-1
-
+# 20 Jun 2022, 3d-gussner, Change to Ardunio_boards v1.0.5-1
+# 20 Jun 2022, wavexx    , New PO-based language translation support
+# 20 Jun 2022, 3d-gussner, fix Mk2.5/S zip after change PO-based language translation support
+# 06 Jul 2022, 3d-gussner, Change to v1.0.8 and Ardunio_boards v1.0.5-2
+# 06 Jul 2022, 3d-gussner, Fix branch check
+# 12 Jul 2022, 3d-gussner, Check if FW_FLAVAVOR and FW_FLAVERSION are correct
 
 SCRIPT_PATH="$( cd "$(dirname "$0")" ; pwd -P )"
+export SRCDIR=$SCRIPT_PATH
 
 #### Start: Failures
 failures()
@@ -189,10 +202,11 @@ case "$1" in
     12) echo "$(tput setaf 5)Failed to copy file $(tput sgr0)" ; exit 12 ;;
     13) echo "$(tput setaf 5)Failed to delete $(tput sgr0)" ; exit 13 ;;
     20) echo "$(tput setaf 2)Conditional stop initiated by user $(tput sgr0)" ; exit 20 ;;
-    21) echo "$(tput setaf 1)PF-build.sh has been interrupted/failed. $(tput setaf 6)Restoring 'Configuration.h'$(tput sgr0)" ; sleep 5 ;;
-    22) echo "$(tput setaf 1)PF-build.sh has been interrupted/failed. $(tput setaf 6)Restoring 'config.h'$(tput sgr0)" ; sleep 5 ;;
+    21) echo "$(tput setaf 1)PF-build.sh has been interrupted/failed. $(tput setaf 6)Restoring 'Configuration.h'$(tput sgr0)" ; if [ $OUTPUT == "1" ] ; then sleep 5 ; fi ;;
+    22) echo "$(tput setaf 1)PF-build.sh has been interrupted/failed. $(tput setaf 6)Restoring 'config.h'$(tput sgr0)" ; if [ $OUTPUT == "1" ] ; then sleep 5 ; fi ;;
     24) echo "$(tput setaf 1)PF-build.sh stopped due to compiling errors! Try to restore modified files.$(tput sgr0)"; check_script_failed_nr1 ; check_script_failed_nr2 ; cleanup_firmware ; exit 24 ;;
     25) echo "$(tput setaf 1)Failed to execute $(tput sgr0)" ; exit 25 ;;
+    26) echo "$(tput setaf 1)FW_FLAVAVOR or FW_FLAVERSION commented out in 'Firmware/Configuration.h'$(tput sgr0)"; check_script_failed_nr1 ; check_script_failed_nr2 ; cleanup_firmware ; exit 26 ;;
 esac
 }
 #### End: Failures
@@ -226,7 +240,7 @@ while getopts b:c:d:g:h:i:j:l:m:n:o:p:v:x:y:?h flag
 # '?' 'h' argument usage and help
 if [ "$help_flag" == "1" ] ; then
 echo "***************************************"
-echo "* PF-build.sh Version: 2.0.2-Build_69 *"
+echo "* PF-build.sh Version: 2.0.2-Build_80 *"
 echo "***************************************"
 echo "Arguments:"
 echo "$(tput setaf 2)-b$(tput sgr0) Build/commit number"
@@ -533,11 +547,11 @@ fi
 #### Start: Set build environment 
 set_build_env_variables()
 {
-BUILD_ENV="1.0.7"
+BUILD_ENV="1.0.8"
 BOARD="prusa_einsy_rambo"
 BOARD_PACKAGE_NAME="PrusaResearch"
 if [ "$ARDUINO_ENV" == "1.8.19" ]; then
-    BOARD_VERSION="1.0.5-1"
+    BOARD_VERSION="1.0.5-2"
 else
     BOARD_VERSION="1.0.4"
 fi
@@ -827,7 +841,8 @@ if [ -z "$variant_flag" ] ; then
     while IFS= read -r -d $'\0' f; do
         options[i++]="$f"
     done < <(find Firmware/variants/ -maxdepth 1 -type f -name "*.h" -print0 )
-    select opt in "${options[@]}" "All" "Quit"; do
+    IFS=$'\n' sorted=($(sort -n <<<"${options[*]}")); unset IFS
+    select opt in "${sorted[@]}" "All" "Quit"; do
         case $opt in
             *.h)
                 VARIANT=$(basename "$opt" ".h")
@@ -915,7 +930,6 @@ fi
 #Check if Build is selected via argument '-b'
 if [ ! -z "$build_flag" ] ; then
     if [[ "$build_flag" == "Auto" && "$git_available" == "1" ]] ; then
-        echo "Build changed to $build_flag"
         BUILD=$(git rev-list --count HEAD)
     elif [[ $build_flag =~ ^[0-9]+$ ]] ; then
         BUILD=$build_flag
@@ -924,27 +938,26 @@ if [ ! -z "$build_flag" ] ; then
         echo "Only $(tput setaf 2)'Auto' (git needed) or numbers $(tput sgr0) are allowed as build '-b' argument!$(tput sgr0)"
         failures 5
     fi
-    echo "New Build number is: $BUILD"
 fi
 
 #Check git branch has changed
 if [ ! -z "git_available" ]; then
-    BRANCH=""
-    CLEAN_PF_FW_BUILD=0
-else
     BRANCH=$(git branch --show-current)
     echo "Current branch is:" $BRANCH
     if [ ! -f "$SCRIPT_PATH/../PF-build.branch" ]; then
-        echo "$BRANCH" >| $SCRIPT_PATH/../PF-build.branch
+        #echo "$BRANCH" >| $SCRIPT_PATH/../PF-build.branch
         echo "created PF-build.branch file"
     else
         PRE_BRANCH=$(cat "$SCRIPT_PATH/../PF-build.branch")
-        echo "Previous branch was:" $PRE_BRANCH
+        #echo "Previous branch was:" $PRE_BRANCH
         if [ ! "$BRANCH" == "$PRE_BRANCH" ] ; then
             CLEAN_PF_FW_BUILD=1
             echo "$BRANCH" >| $SCRIPT_PATH/../PF-build.branch
         fi
     fi
+else
+    BRANCH=""
+    CLEAN_PF_FW_BUILD=0
 fi
 }
 #### End: Getting arguments for command line compiling
@@ -1008,20 +1021,27 @@ prepare_code_for_compiling()
     MOTHERBOARD=$(grep --max-count=1 "\bMOTHERBOARD\b" $SCRIPT_PATH/Firmware/variants/$VARIANT.h | sed -e's/  */ /g' |cut -d ' ' -f3)
     # Check development status
     FW_FLAV=$(grep --max-count=1 "//#define FW_FLAVOR\b" $SCRIPT_PATH/Firmware/Configuration.h|cut -d ' ' -f1)
+    FW_FLAVREV=$(grep --max-count=1 "//#define FW_FLAVERSION\b" $SCRIPT_PATH/Firmware/Configuration.h|cut -d ' ' -f1)
+    if [[ "$FW_FLAV" == "//#define" && -z $FW_FLAVREV ]];then
+        failures 26
+    fi
     #echo "FLAV:$FW_FLAV"
     if [[ "$FW_FLAV" != "//#define" ]] ; then
         FW_FLAVOR=$(grep --max-count=1 "\bFW_FLAVOR\b" $SCRIPT_PATH/Firmware/Configuration.h| sed -e's/  */ /g'|cut -d ' ' -f3)
         #echo "FLAVOR:$FW_FLAVOR"
-        FW_FLAVERSION=$(grep --max-count=1 "\bFW_FLAVERSION\b" $SCRIPT_PATH/Firmware/Configuration.h| sed -e's/  */ /g'|cut -d ' ' -f3)
-        if [[ "$FW_FLAVOR" != "//#define FW_FLAVOR" ]] ; then
+        if [[ "$FW_FLAVREV" != "//#define" ]] ; then
+            FW_FLAVERSION=$(grep --max-count=1 "\bFW_FLAVERSION\b" $SCRIPT_PATH/Firmware/Configuration.h| sed -e's/  */ /g'|cut -d ' ' -f3)
             FW="$FW-$FW_FLAVOR"
             DEV_CHECK="$FW_FLAVOR"
-            #echo "DEV:$DEV_CHECK"
+            echo "DEV:$DEV_CHECK"
             if [ ! -z "$FW_FLAVERSION" ] ; then
                 FW="$FW$FW_FLAVERSION"
             fi
+        else
+            failures 26
         fi
     fi
+    #DEV_CHECK=$(grep --max-count=1 "\bFW_VERSION\b" $SCRIPT_PATH/Firmware/Configuration.h | sed -e's/  */ /g'|cut -d '"' -f2|sed 's/\.//g'|cut -d '-' -f2)
     if [ -z "$DEV_STATUS_SELECTED" ] ; then
         if [[ "$DEV_CHECK" == *"RC"* ]] ; then
             DEV_STATUS="RC"
@@ -1145,7 +1165,7 @@ prepare_variant_for_compiling()
     fi
 
     #Prepare Configuration.h to use the correct FW_DEV_VERSION to prevent LCD messages when connecting with OctoPrint
-    sed -i -- "s/#define FW_DEV_VERSION FW_VERSION_UNKNOWN/#define FW_DEV_VERSION FW_VERSION_$DEV_STATUS/g" $SCRIPT_PATH/Firmware/Configuration.h
+    sed -i -- "s/#define FW_DEV_VERSION FW_VERSION_.*/#define FW_DEV_VERSION FW_VERSION_$DEV_STATUS/g" $SCRIPT_PATH/Firmware/Configuration.h
 
     # set FW_REPOSITORY
     sed -i -- 's/#define FW_REPOSITORY "Unknown"/#define FW_REPOSITORY "Prusa3d"/g' $SCRIPT_PATH/Firmware/Configuration.h
@@ -1321,15 +1341,9 @@ create_multi_firmware()
             ./fw-clean.sh
             echo "$(tput sgr 0)"
         fi
-        # build languages
+        # Combine compiled firmware with languages
         echo "$(tput setaf 3)"
-        ./lang-build.sh || failures 25
-        # build community languages
-        ./lang-community.sh || failures 25
-        # Combine compiled firmware with languages 
         ./fw-build.sh || failures 25
-        cp not_tran.txt not_tran_$VARIANT.txt
-        cp not_used.txt not_used_$VARIANT.txt
         echo "$(tput sgr 0)"
         # Check if the motherboard is an EINSY and if so only one hex file will generated
         MOTHERBOARD=$(grep --max-count=1 "\bMOTHERBOARD\b" $SCRIPT_PATH/Firmware/variants/$VARIANT.h | sed -e's/  */ /g' |cut -d ' ' -f3)
@@ -1337,28 +1351,22 @@ create_multi_firmware()
         if [ "$MOTHERBOARD" = "BOARD_EINSY_1_0a" ]; then
             echo "$(tput setaf 2)Copying multi language firmware for MK3/Einsy board to PF-build-hex folder$(tput sgr 0)"
             # End of "lang.bin" for MK3 and MK3S copy
-            cp -f firmware.hex $SCRIPT_PATH/../$OUTPUT_FOLDER/$OUTPUT_FILENAME.hex
+            cp -f Firmware-intl.hex $SCRIPT_PATH/../$OUTPUT_FOLDER/$OUTPUT_FILENAME.hex
             cp -f $BUILD_PATH/Firmware.ino.elf $SCRIPT_PATH/../$OUTPUT_FOLDER/$OUTPUT_FILENAME.elf
         else
-            echo "$(tput setaf 2)Zip multi language firmware for MK2.5/miniRAMbo board to PF-build-hex folder$(tput sgr 0)"
-            cp -f firmware_cz.hex $SCRIPT_PATH/../$OUTPUT_FOLDER/$OUTPUT_FILENAME-cz.hex
-            cp -f firmware_de.hex $SCRIPT_PATH/../$OUTPUT_FOLDER/$OUTPUT_FILENAME-de.hex
-            cp -f firmware_es.hex $SCRIPT_PATH/../$OUTPUT_FOLDER/$OUTPUT_FILENAME-es.hex
-            cp -f firmware_fr.hex $SCRIPT_PATH/../$OUTPUT_FOLDER/$OUTPUT_FILENAME-fr.hex
-            cp -f firmware_it.hex $SCRIPT_PATH/../$OUTPUT_FOLDER/$OUTPUT_FILENAME-it.hex
-            cp -f firmware_pl.hex $SCRIPT_PATH/../$OUTPUT_FOLDER/$OUTPUT_FILENAME-pl.hex
-            cp -f firmware_nl.hex $SCRIPT_PATH/../$OUTPUT_FOLDER/$OUTPUT_FILENAME-nl.hex
+            #Search for created firmware languages
+            langs=$(find Firmware-intl-en_*.hex | cut -d "_" -f2 | cut -d "." -f1)
+            #Copy found firmware_*.hex files 
+                for la in $langs; do
+                    cp -f Firmware-intl-en_$la.hex $SCRIPT_PATH/../$OUTPUT_FOLDER/$OUTPUT_FILENAME-$la.hex
+                done
             cp -f $BUILD_PATH/Firmware.ino.elf $SCRIPT_PATH/../$OUTPUT_FOLDER/$OUTPUT_FILENAME.elf
+            echo "$(tput setaf 2)Zip multi language firmware for MK2.5/miniRAMbo board to PF-build-hex folder$(tput sgr 0)"
             if [ $TARGET_OS == "windows" ]; then 
                 zip a $SCRIPT_PATH/../$OUTPUT_FOLDER/$OUTPUT_FILENAME.zip $SCRIPT_PATH/../$OUTPUT_FOLDER/$OUTPUT_FILENAME-??.hex
-                rm $SCRIPT_PATH/../$OUTPUT_FOLDER/$OUTPUT_FILENAME-??.hex
+                #rm $SCRIPT_PATH/../$OUTPUT_FOLDER/$OUTPUT_FILENAME-??.hex
             elif [ $TARGET_OS == "linux" ]; then
-                # Make a copy for MK404 sim of MK2, MK2.5, MK2.5S firmware
-                if [ ! -z "$mk404_flag" ]; then
-                    cp -f firmware_de.hex $SCRIPT_PATH/../$OUTPUT_FOLDER/$OUTPUT_FILENAME.hex
-                fi
-                # End of MK2, MK2.5, MK2.5S firmware copy
-            zip -m -j ../../$OUTPUT_FOLDER/$OUTPUT_FILENAME.zip ../../$OUTPUT_FOLDER/$OUTPUT_FILENAME-??.hex
+                zip -j ../../$OUTPUT_FOLDER/$OUTPUT_FILENAME.zip ../../$OUTPUT_FOLDER/$OUTPUT_FILENAME-??.hex
             fi
         fi
 
@@ -1366,7 +1374,6 @@ create_multi_firmware()
     if [[ -z "$clean_flag" || "$clean_flag" == "0" ]]; then
         echo "$(tput setaf 3)"
         ./fw-clean.sh || failures 25
-        ./lang-clean.sh || failures 25
         echo "$(tput sgr 0)"
     fi
 }
@@ -1388,7 +1395,9 @@ save_en_firmware()
 cleanup_firmware()
 {
     if [[ -z "$prusa_flag" || "$prusa_flag" == "0" ]]; then
-        rm $SCRIPT_PATH/Firmware/Configuration_prusa.h || failures 13
+        if [ -e "$SCRIPT_PATH/Firmware/Configuration_prusa.h" ]; then
+            rm $SCRIPT_PATH/Firmware/Configuration_prusa.h || failures 13
+        fi
     fi
     # Delete dupblicates
     if find $SCRIPT_PATH/lang/ -name '*RAMBo10a*.txt' -printf 1 -quit | grep -q 1
@@ -1415,8 +1424,8 @@ cleanup_firmware()
     fi
 
     # Restore files to previous state
-    sed -i -- "s/^#define FW_DEV_VERSION FW_VERSION_$DEV_STATUS/#define FW_DEV_VERSION FW_VERSION_UNKNOWN/g" $SCRIPT_PATH/Firmware/Configuration.h
-    sed -i -- 's/^#define FW_REPOSITORY "Prusa3d"/#define FW_REPOSITORY "Unknown"/g' $SCRIPT_PATH/Firmware/Configuration.h
+    sed -i -- "s/^#define FW_DEV_VERSION FW_VERSION_.*/#define FW_DEV_VERSION FW_VERSION_UNKNOWN/g" $SCRIPT_PATH/Firmware/Configuration.h
+    sed -i -- 's/^#define FW_REPOSITORY.*/#define FW_REPOSITORY "Unknown"/g' $SCRIPT_PATH/Firmware/Configuration.h
     if [ ! -z "$BUILD_ORG" ] ; then
         sed -i -- "s/^#define FW_COMMIT_NR.*/#define FW_COMMIT_NR $BUILD_ORG/g" $SCRIPT_PATH/Firmware/Configuration.h
     fi
@@ -1440,11 +1449,11 @@ cleanup_firmware()
     fi
 
     # Restore build env files to previous state
-    if [ $BOARD_MEM_MODIFIED == "1" ]; then
+    if [ "$BOARD_MEM_MODIFIED" == "1" ]; then
         sed -i -- "s/^#define FLASHEND .*$/#define FLASHEND        0x3FFFF/g" $BUILD_ENV_PATH/hardware/tools/avr/avr/include/avr/iom2560.h
         echo "$(tput setaf 2)Restored Board Mem$(tput sgr 0)"
     fi
-    if [ $BOARD_FLASH_MODIFIED == "1" ]; then
+    if [ "$BOARD_FLASH_MODIFIED" == "1" ]; then
         sed -i -- "s/^prusa_einsy_rambo.upload.maximum_size.*/prusa_einsy_rambo.upload.maximum_size=253952/g" $BUILD_ENV_PATH/portable/packages/$BOARD_PACKAGE_NAME/hardware/avr/$BOARD_VERSION/boards.txt
         sed -i -- "s/^#define RAMEND.*/#define RAMEND          0x21FF/g" $BUILD_ENV_PATH/hardware/tools/avr/avr/include/avr/iom2560.h
         echo "$(tput setaf 2)Restored Board Flash$(tput sgr 0)"
@@ -1482,7 +1491,7 @@ if [[ "$output_flag" == "1" || -z "$output_flag" ]]; then
     if [[ -z "$mk404_flag" && "$variant_flag" != "All" ]]; then
         echo
         read -t 10 -n 1 -p "Do you want to start MK404? Y/$(tput setaf 2)n$(tput sgr 0)" mk404_start
-        if [ "$mk404_start" == "Y" ]; then
+        if [[ "$mk404_start" == "Y" || "$mk404_start" == "y" ]]; then
             echo
             read -t 10 -n 1 -p "Do you want to start MK404 with or without MMU2S? $(tput setaf 2)1$(tput sgr 0)/2" mk404_choose1
             if [ "$mk404_choose1" == "1" ]; then
@@ -1525,11 +1534,6 @@ fi
 
 if [[ ! -z "$mk404_flag" && "$variant_flag" != "All " ]]; then
 
-# For Prusa MK2, MK2.5/S
-    if [ "$MOTHERBOARD" == "BOARD_RAMBO_MINI_1_3" ]; then
-        MK404_PRINTER="${MK404_PRINTER}_mR13"
-    fi
-
 # Run MK404 with 'debugcore' and/or 'bootloader-file'
     if [ ! -z "$board_mem_flag" ]; then
         MK404_options="-x $board_mem_flag"
@@ -1560,7 +1564,16 @@ if [[ ! -z "$mk404_flag" && "$variant_flag" != "All " ]]; then
 
 #Decide which hex file to use EN_ONLY or Multi language
     if [ "$LANGUAGES" == "ALL" ]; then
-        MK404_firmware_file=$SCRIPT_PATH/../$OUTPUT_FOLDER/$OUTPUT_FILENAME.hex
+        if [[ "$MK404_PRINTER" == "MK3" || "$MK404_PRINTER" == "MK3S" ]]; then
+            MK404_firmware_file=$SCRIPT_PATH/../$OUTPUT_FOLDER/$OUTPUT_FILENAME.hex
+        else
+            PS3="Select a language:"
+            select lan in ${langs[@]}
+            do
+                MK404_firmware_file=$SCRIPT_PATH/../$OUTPUT_FOLDER/$OUTPUT_FILENAME-$lan.hex
+                break
+            done
+        fi
     else
         MK404_firmware_file=$SCRIPT_PATH/../$OUTPUT_FOLDER/$OUTPUT_FILENAME-EN_ONLY.hex
     fi
